# dashboard/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: MDDash Dashboard API
  version: 1.0.0
  description: API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
servers:
  - url: /
...
paths:
...
  /dash/api/experiments:
    get:
      tags: [Experiments]
      summary: List experiments
      operationId: listExperiments
      responses:
        '200':
          description: Experiments
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Experiment'}}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Experiments]
      summary: Create an experiment
      operationId: createExperiment
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: {$ref: '#/components/schemas/CreateExperimentForm'}
      responses:
        '201':
          description: Created experiment
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Experiment'}
        '400': {$ref: '#/components/responses/BadRequest'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
    get:
      tags: [Experiments]
      summary: Get an experiment
      operationId: getExperiment
      responses:
        '200':
          description: Experiment
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Experiment'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
    patch:
      tags: [Experiments]
      summary: Update an experiment
      operationId: updateExperiment
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateExperiment'}
      responses:
        '200':
          description: Updated experiment
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Experiment'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Experiments]
      summary: Delete an experiment
      operationId: deleteExperiment
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/publish:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
    post:
      tags: [Experiments]
      summary: Publish an experiment
      operationId: publishExperiment
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PublishRequest'}
      responses:
        '201':
          description: MDPosit publication package
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MDPositPublication'}
        '202':
          description: MDRepo publication accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MDRepoPublication'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/publish/status:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
    get:
      tags: [Experiments]
      summary: Get MDRepo upload status
      operationId: getPublishStatus
      responses:
        '200':
          description: Publication status
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PublishStatus'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/mdrepo/status:
    get:
      tags: [MDRepo]
      summary: Get MDRepo authentication status
      operationId: getMDRepoStatus
      responses:
        '200':
          description: Authentication status
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MDRepoStatus'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/mdrepo/auth:
    get:
      tags: [MDRepo]
      summary: Begin MDRepo OAuth authorization
      operationId: authorizeMDRepo
      parameters:
        - name: return_url
          in: query
          schema: {type: string, default: /}
      responses:
        '302': {$ref: '#/components/responses/Redirect'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/mdrepo/callback:
    get:
      tags: [MDRepo]
      summary: Complete MDRepo OAuth authorization
      operationId: completeMDRepoAuthorization
      parameters:
        - {name: code, in: query, schema: {type: string}}
        - {name: state, in: query, schema: {type: string}}
        - {name: error, in: query, schema: {type: string}}
        - {name: error_description, in: query, schema: {type: string}}
      responses:
        '302': {$ref: '#/components/responses/Redirect'}
  /dash/api/mdrepo/logout:
    post:
      tags: [MDRepo]
      summary: Clear MDRepo credentials
      operationId: logoutMDRepo
      responses:
        '200':
          description: Logout confirmation
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Message'}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  parameters:
    ExperimentId: {name: experiment_id, in: path, required: true, schema: {type: string, minLength: 1}}
...
  responses:
...
    NoContent: {description: Operation completed successfully}
    Redirect:
      description: Redirect
      headers:
        Location: {required: true, schema: {type: string}}
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    ServiceUnavailable:
      description: Service unavailable
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details. The HTTP status is carried by the response status line.
      required: [type, title, detail]
      properties:
        type: {type: string, format: uri, pattern: '^urn:mddash:'}
        title: {type: string, minLength: 1}
        detail: {type: string, minLength: 1}
        solution: {type: string, minLength: 1}
      additionalProperties: false
    Engine: {type: string, enum: [GMX, AMBER]}
    JobStatus: {type: string, enum: [UNKNOWN, PENDING, RUNNING, FINISHED, ERROR]}
    PodStatus: {type: string, enum: [UNKNOWN, PENDING, RUNNING, ERROR, DOWN, TERMINATING, TERMINATED]}
    NullableString: {type: [string, 'null']}
    Message:
      type: object
      required: [message]
      properties: {message: {type: string}}
      additionalProperties: false
...
    Notebook:
      type: object
      required: [id, experiment_id, token, gpu, path, status, started_at]
      properties:
        id: {type: integer}
        experiment_id: {type: string}
        token: {type: string}
        tier: {type: [string, 'null'], enum: [1x, 2x, 4x, null]}
        gpu: {type: boolean}
        path: {type: string}
        status: {$ref: '#/components/schemas/PodStatus'}
        started_at: {type: [string, 'null'], format: date-time, description: Pod start time; null while the pod is absent}
      additionalProperties: false
    Experiment:
      type: object
      required: [id, created_at, updated_at, name, engine, latest_simulation_path, notebook, tuner_jobs, simulation_jobs, analysis_jobs]
      properties:
        id: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        name: {type: string}
        module_name: {$ref: '#/components/schemas/NullableString'}
        # Manifest path of the most recently active simulation; the experiment's
        # step/status and dashboard detail rows always inherit from it.
        latest_simulation_path: {type: [string, 'null']}
        source:
          oneOf:
            - {$ref: '#/components/schemas/ExperimentSource'}
            - {type: 'null'}
        size_bytes: {type: [integer, 'null']}
        notebooks_repo: {$ref: '#/components/schemas/NullableString'}
        mdrepo_id: {$ref: '#/components/schemas/NullableString'}
        mdrepo_published: {type: [boolean, 'null']}
        engine: {$ref: '#/components/schemas/Engine'}
        notebook:
          oneOf:
            - {$ref: '#/components/schemas/Notebook'}
            - {type: 'null'}
        tuner_jobs: {type: array, items: {$ref: '#/components/schemas/TunerJob'}}
        simulation_jobs: {type: array, items: {$ref: '#/components/schemas/SimulationJob'}}
        analysis_jobs: {type: array, items: {$ref: '#/components/schemas/AnalysisJob'}}
        step: {type: integer}
        status: {type: string}
        # Publishing is experiment-level: true once a draft exists or any run finished.
        can_publish: {type: boolean}
        mdrepo_record_url: {$ref: '#/components/schemas/NullableString'}
      additionalProperties: false
    CreateExperimentForm:
      type: object
      required: [experiment-name, type]
      properties:
        experiment-name: {type: string, minLength: 1}
        type: {type: string, enum: [pdb, repo, file]}
        pdb: {type: string}
        repo-url: {type: string}
        notebooks-repo: {type: string}
        access-token: {type: string}
        notebook-module: {type: string}
        engine: {$ref: '#/components/schemas/Engine'}
        simulation-files: {type: array, items: {type: string, contentMediaType: application/octet-stream}}
      additionalProperties: false
    UpdateExperiment:
      type: object
      required: [name]
      properties: {name: {type: string}}
      additionalProperties: false
    ExperimentSource:
      type: object
      required: [type, files]
      properties:
        type: {type: string, enum: [pdb, repo, file]}
        pdb_id: {type: string}
        url: {type: string}
        files: {type: array, items: {$ref: '#/components/schemas/FileInfo'}}
        file_count: {type: integer, minimum: 0}
      additionalProperties: false
    FileInfo:
      type: object
      required: [name, size, path, url]
      properties:
        name: {type: string}
        size: {type: integer, minimum: 0}
        path: {type: string}
        url: {type: string}
      additionalProperties: false
...
    SimulationJob:
      type: object
      required: [id, experiment_id, simulation_path, created_at, engine, np, ntomp, status, is_live]
      properties:
        id: {type: string}
        experiment_id: {type: string}
        simulation_path: {type: string}
        created_at: {type: string, format: date-time}
        engine: {$ref: '#/components/schemas/Engine'}
        np: {type: integer, minimum: 1}
        ntomp: {type: integer, minimum: 1}
        status: {$ref: '#/components/schemas/JobStatus'}
        start_timestamp: {type: [integer, 'null']}
        finish_timestamp: {type: [integer, 'null']}
        nsteps: {type: [integer, 'null']}
        nsteps_done: {type: [integer, 'null'], description: Steps completed so far (parsed from the engine log)}
        performance: {type: [number, 'null']}
        estimated_time: {type: [integer, 'null'], description: Estimated seconds until completion}
        # True while the job's status is non-terminal; auto-emitted with every job payload.
        is_live: {type: boolean}
        log_lines:
          type: object
          description: Line counts per log stream for sizing badges without fetching logs; null while a stream's file does not exist yet
          additionalProperties: false
          properties:
            gmx: {type: [integer, 'null'], description: Engine log (GROMACS only)}
            mdout: {type: [integer, 'null'], description: Engine log (AMBER only)}
            stdout: {type: [integer, 'null']}
            stderr: {type: [integer, 'null']}
      additionalProperties: true
...
    TunerTrial:
      type: object
      properties:
        id: {type: string}
        status: {$ref: '#/components/schemas/JobStatus'}
        performance: {type: [number, 'null']}
      additionalProperties: true
    TunerJob:
      type: object
      required: [id, experiment_id, simulation_path, created_at, nsteps, is_stopped, engine, tuner_status, trials, is_live]
      properties:
        id: {type: string}
        experiment_id: {type: string}
        simulation_path: {type: string}
        nsteps: {type: integer}
        error_message: {$ref: '#/components/schemas/NullableString'}
        created_at: {type: string, format: date-time}
        is_stopped: {type: boolean}
        engine: {$ref: '#/components/schemas/Engine'}
        tuner_status: {$ref: '#/components/schemas/JobStatus'}
        # True while the tuner job is not stopped and its status is non-terminal.
        is_live: {type: boolean}
        sim_length_ns: {type: [number, 'null']}
        trials: {type: array, items: {$ref: '#/components/schemas/TunerTrial'}}
      additionalProperties: false
...
    AnalysisJob:
      type: object
      required: [id, experiment_id, simulation_path, analysis_name, created_at, status]
      properties:
        id: {type: string}
        experiment_id: {type: string}
        simulation_path: {type: string}
        analysis_name:
          type: string
          enum: [apl, clusters, density, dist, energies, hbonds, inter, linter, lorder, pairwise, pca, perres, pockets, rgyr, rmsds, rmsf, sas, thickness, tmscore]
        created_at: {type: string, format: date-time}
        status: {$ref: '#/components/schemas/JobStatus'}
        sim_progress:
          type: number
          nullable: true
          description: Fraction (0-1) of the simulation's steps available when the analysis inputs were snapshotted; null when unknown.
      additionalProperties: true
...
    PublishRequest:
      type: object
      properties:
        target: {type: string, enum: [invenio, mdposit], default: invenio}
        simulation_path: {type: [string, 'null']}
      additionalProperties: false
    MDRepoPublication:
      type: object
      required: [id, links, upload_id, upload_state, draft_url]
      properties:
        id: {type: string}
        links: {type: object, additionalProperties: {type: string}}
        upload_id: {type: string}
        upload_state: {type: string}
        draft_url: {$ref: '#/components/schemas/NullableString'}
      additionalProperties: false
    PublicationFile:
      type: object
      required: [path, url]
      properties:
        role: {type: string}
        path: {type: string}
        url: {type: string}
      additionalProperties: false
    MDPositPublication:
      type: object
      required: [metadata_file, files, vre_lite_url]
      properties:
        metadata_file: {$ref: '#/components/schemas/PublicationFile'}
        files: {type: array, items: {$ref: '#/components/schemas/PublicationFile'}}
        vre_lite_url: {$ref: '#/components/schemas/NullableString'}
      additionalProperties: false
    PublishStatus:
      type: object
      required: [experiment_id, mdrepo_id, draft_url, upload_state, reason, total_files, completed_files, total_bytes, completed_bytes]
      properties:
        experiment_id: {type: string}
        mdrepo_id: {$ref: '#/components/schemas/NullableString'}
        draft_url: {$ref: '#/components/schemas/NullableString'}
        upload_state: {type: [string, 'null']}
        reason: {$ref: '#/components/schemas/NullableString'}
        total_files: {type: integer, minimum: 0}
        completed_files: {type: integer, minimum: 0}
        total_bytes: {type: integer, minimum: 0}
        completed_bytes: {type: integer, minimum: 0}
        current_file: {$ref: '#/components/schemas/NullableString'}
        error: {$ref: '#/components/schemas/NullableString'}
        upload_attempt_id: {type: string}
        failed_files: {type: array, items: {$ref: '#/components/schemas/FailedUploadFile'}}
      additionalProperties: true
    FailedUploadFile:
      type: object
      required: [key, error]
      properties:
        key: {type: string}
        error: {type: string}
      additionalProperties: true
    MDRepoStatus:
      type: object
      required: [authenticated]
      properties:
        authenticated: {type: boolean}
        mdrepo_url: {type: string}
      additionalProperties: false
...