# dashboard/api/routes/mdrepo.py
"""
MDRepo OAuth2 routes for authentication with the InvenioRDM repository.

This module handles the OAuth2 authorization code flow to obtain access tokens
for the MDRepo API.
"""

import logging
import secrets
import time
from http import HTTPStatus
from urllib.parse import parse_qsl, urlencode, urlsplit, urlunsplit

import requests
from config import (
    API_PREFIX,
    MDREPO_AUTHORIZE_URL,
    MDREPO_CLIENT_ID,
    MDREPO_CLIENT_SECRET,
    MDREPO_REDIRECT_URI,
    MDREPO_SCOPES,
    MDREPO_TOKEN_URL,
    MDREPO_URL,
)
from errors import ApiError
from flask import Blueprint, Response, jsonify, redirect, request, session
from models.experiment import mdrepo_status_cache
from token_manager import (
    MDREPO_REFRESH_TOKEN_KEY,
    MDREPO_STATE_KEY,
    MDREPO_TOKEN_EXPIRES_AT,
    MDREPO_TOKEN_KEY,
    MDRepoTokenManager,
)
from werkzeug.wrappers import Response as WerkzeugResponse

logger = logging.getLogger(__name__)

mdrepo_bp = Blueprint("mdrepo", __name__, url_prefix=f"{API_PREFIX}/mdrepo")


def _redirect_with_query(url: str, **params: str) -> WerkzeugResponse:
    """Redirect to a URL with query params merged in — the return URL may already carry its own query string."""
    parts = urlsplit(url)
    query = dict(parse_qsl(parts.query, keep_blank_values=True))
    query.update(params)
    return redirect(urlunsplit((parts.scheme, parts.netloc, parts.path, urlencode(query), parts.fragment)))


def _safe_return_url(url: str) -> str:
    r"""Same-origin paths only (open-redirect guard, CWE-601): absolute, protocol-relative, and backslash variants fall back to ``/``."""
    parts = urlsplit(url)
    if parts.scheme or parts.netloc or "\\" in url or not url.startswith("/") or url.startswith("//"):
        return "/"
    return url


def get_mdrepo_token() -> str | None:
    """
    Get the MDRepo access token from the current session.

    Returns:
        The access token string if present in the session, or None.
    """
    return session.get(MDREPO_TOKEN_KEY)


@mdrepo_bp.route("/status", methods=["GET"])
def get_status() -> Response:
    """
    Check if user has a valid MDRepo token.

    Returns:
        Response: JSON response with authentication status and MDRepo URL if authenticated.
    """
    token = session.get(MDREPO_TOKEN_KEY)

    if not token:
        return jsonify({"authenticated": False})

    try:
        resp = requests.get(f"{MDREPO_URL}/api/me", headers={"Authorization": f"Bearer {token}"}, timeout=10)

        if resp.status_code == HTTPStatus.OK:
            return jsonify({"authenticated": True, "mdrepo_url": MDREPO_URL})

        # Token invalid or expired
        session.pop(MDREPO_TOKEN_KEY, None)
        return jsonify({"authenticated": False})

    except Exception as e:
        logger.error(f"MDRepo token validation failed: {e}")
        return jsonify({"authenticated": False})


@mdrepo_bp.route("/auth", methods=["GET"])
def initiate_auth() -> Response | WerkzeugResponse:
    """
    Initiate OAuth2 authorization flow with MDRepo.

    Query params:
        return_url: URL to redirect to after successful authentication.

    Returns:
        Response: Redirect to the MDRepo authorization URL, or a JSON error if OAuth is not configured.
    """
    if not all([MDREPO_CLIENT_ID, MDREPO_CLIENT_SECRET, MDREPO_REDIRECT_URI]):
        raise ApiError(
            HTTPStatus.SERVICE_UNAVAILABLE,
            "MDRepo OAuth is not configured. Contact administrator.",
            "urn:mddash:mdrepo-not-configured",
            "MDRepo integration isn't enabled on this deployment; contact the administrator.",
        )

    # The stored return URL becomes the post-OAuth 302 target.
    return_url = _safe_return_url(request.args.get("return_url", "/"))
    state = secrets.token_urlsafe(32)
    session[MDREPO_STATE_KEY] = state
    session["mdrepo_return_url"] = return_url

    # Build authorization URL
    params = {
        "client_id": MDREPO_CLIENT_ID,
        "redirect_uri": MDREPO_REDIRECT_URI,
        "response_type": "code",
        "state": state,
        "scope": MDREPO_SCOPES,
    }

    auth_url = f"{MDREPO_AUTHORIZE_URL}?{urlencode(params)}"
    logger.info(f"Redirecting to MDRepo OAuth: {MDREPO_AUTHORIZE_URL}")

    return redirect(auth_url)


@mdrepo_bp.route("/callback", methods=["GET"])
def oauth_callback() -> WerkzeugResponse:
    """
    Handle OAuth2 callback from MDRepo.

    Returns:
        WerkzeugResponse: Redirect to the return URL with a success or error query parameter.
    """
    code = request.args.get("code")
    state = request.args.get("state")
    error = request.args.get("error")

    return_url = session.pop("mdrepo_return_url", "/")
    stored_state = session.pop(MDREPO_STATE_KEY, None)

    # Check for OAuth errors
    if error:
        error_description = request.args.get("error_description", "Authorization denied")
        logger.error(f"MDRepo OAuth error: {error} - {error_description}")
        return _redirect_with_query(return_url, mdrepo_error=error_description)

    # Validate state to prevent CSRF
    if not state or state != stored_state:
        logger.error("MDRepo OAuth: Invalid state parameter")
        return _redirect_with_query(return_url, mdrepo_error="Invalid state parameter")

    if not code:
        logger.error("MDRepo OAuth: No authorization code received")
        return _redirect_with_query(return_url, mdrepo_error="No authorization code")

    # Exchange code for token
    try:
        token_data = {
            "grant_type": "authorization_code",
            "code": code,
            "redirect_uri": MDREPO_REDIRECT_URI,
            "client_id": MDREPO_CLIENT_ID,
            "client_secret": MDREPO_CLIENT_SECRET,
        }

        response = requests.post(MDREPO_TOKEN_URL, data=token_data, timeout=30)

        if not response.ok:
            logger.error(f"MDRepo token exchange failed: {response.status_code} - {response.text}")
            return _redirect_with_query(return_url, mdrepo_error="Token exchange failed")

        token_response = response.json()
        access_token = token_response.get("access_token")
        refresh_token = token_response.get("refresh_token")
        expires_in = token_response.get("expires_in", 3600)  # Default 1 hour

        if not access_token:
            logger.error("MDRepo OAuth: No access token in response")
            return _redirect_with_query(return_url, mdrepo_error="No access token")

        # Store complete token information in session
        session[MDREPO_TOKEN_KEY] = access_token
        if refresh_token:
            session[MDREPO_REFRESH_TOKEN_KEY] = refresh_token
            logger.info("MDRepo OAuth: Successfully obtained access and refresh tokens")
        else:
            logger.warning("MDRepo OAuth: No refresh token in response - token refresh will not be available")
        session[MDREPO_TOKEN_EXPIRES_AT] = time.time() + expires_in

        # Clear MDRepo status cache to force re-sync with new token
        mdrepo_status_cache.clear()
        logger.info("MDRepo OAuth: Cleared mdrepo_status_cache")

        return _redirect_with_query(return_url, mdrepo_auth="success")

    except requests.RequestException as e:
        logger.error(f"MDRepo OAuth request error: {e}")
        return _redirect_with_query(return_url, mdrepo_error="Request failed")


@mdrepo_bp.route("/logout", methods=["POST"])
def logout() -> Response:
    """
    Remove all MDRepo tokens from session.

    Returns:
        Response: JSON response confirming successful logout.
    """
    token_manager = MDRepoTokenManager(session)
    token_manager.clear_tokens()
    return jsonify({"message": "Logged out from MDRepo"})